% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{calc_geometric_mean}
\alias{calc_geometric_mean}
\title{Calculate Geometric Mean}
\usage{
calc_geometric_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector. All values must be strictly positive.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}
}
\value{
A numeric value.
}
\description{
Computes the geometric mean of a numeric vector. Highly useful for rates of change.
}
