% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{calc_harmonic_mean}
\alias{calc_harmonic_mean}
\title{Calculate Harmonic Mean}
\usage{
calc_harmonic_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector. All values must be strictly positive.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}
}
\value{
A numeric value.
}
\description{
Computes the harmonic mean of a numeric vector.
}
