% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{calc_weighted_mean}
\alias{calc_weighted_mean}
\title{Calculate a Weighted Mean}
\usage{
calc_weighted_mean(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{w}{A numeric vector of weights, the same length as x.}

\item{na.rm}{Logical. Should missing values be removed? Defaults to TRUE.}
}
\value{
A numeric value representing the weighted mean.
}
\description{
Computes the weighted average of a numeric vector.
}
