% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cat_barplot}
\alias{cat_barplot}
\title{Bar Chart for Categorical Data}
\usage{
cat_barplot(x)
}
\arguments{
\item{x}{categorical variable (vector)}
}
\value{
A ggplot object
}
\description{
Bar Chart for Categorical Data
}
