% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catgorical.R
\name{cat_freq}
\alias{cat_freq}
\title{Categorical Frequency Table}
\usage{
cat_freq(x)
}
\arguments{
\item{x}{Character or factor vector}
}
\value{
Data frame
}
\description{
Creates frequency, relative frequency, and cumulative frequency table
for categorical variables.
}
