% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{cat_num_freq}
\alias{cat_num_freq}
\title{Categorical vs Numerical Frequency Table}
\usage{
cat_num_freq(cat_var, num_var)
}
\arguments{
\item{cat_var}{categorical vector}

\item{num_var}{numeric vector}
}
\value{
A data frame with grouped descriptive frequencies
}
\description{
Categorical vs Numerical Frequency Table
}
