% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{data_diagnostics}
\alias{data_diagnostics}
\title{Data Cleansing and Diagnostic Summary}
\usage{
data_diagnostics(x)
}
\arguments{
\item{x}{Target data frame or numeric/character vector}
}
\value{
A list containing the diagnostic summary and the cleaned data
}
\description{
Detects duplicates, removes them, and provides a missing value breakdown.
Now fully supports both data frames and single variables/vectors.
}
