% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catgorical.R
\name{diagnostic_stats}
\alias{diagnostic_stats}
\title{Calculate Diagnostic Test Accuracy Metrics}
\usage{
diagnostic_stats(var1, var2 = NULL)
}
\arguments{
\item{var1}{A 2x2 matrix/table, OR a categorical vector representing the Test Result.}

\item{var2}{A categorical vector representing the Gold Standard / True Condition. Leave NULL if var1 is a table.}
}
\value{
A data frame with estimates and 95\% exact confidence intervals for each metric.
}
\description{
Computes Sensitivity, Specificity, Positive Predictive Value (PPV), and
Negative Predictive Value (NPV) with Exact 95\% Confidence Intervals.
}
\note{
Assumes that the first row/column represents the "Positive" condition and the second represents "Negative".
}
