% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{.ensure_grouped_df}
\alias{.ensure_grouped_df}
\title{Smart Helper: Ensure Grouped Data}
\usage{
.ensure_grouped_df(data)
}
\arguments{
\item{data}{The input numeric vector or frequency table to check.}
}
\value{
A validated frequency data frame.
}
\description{
(Internal function, not exported)
Checks if input is raw data or a table. Converts raw data to a table automatically.
}
