% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{group_boxplot}
\alias{group_boxplot}
\title{Grouped Boxplot}
\usage{
group_boxplot(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A ggplot object
}
\description{
Creates a modern, horizontal boxplot for a numeric vector.
}
