% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_cv}
\alias{group_cv}
\title{Grouped Coefficient of Variation}
\usage{
group_cv(x)
}
\arguments{
\item{x}{Numeric vector OR frequency table}
}
\value{
CV percentage
}
\description{
Grouped Coefficient of Variation
}
