% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{group_desc_stats}
\alias{group_desc_stats}
\title{Grouped Descriptive Statistics}
\usage{
group_desc_stats(data)
}
\arguments{
\item{data}{A numeric vector OR a dataframe output from group_exclusive_freq}
}
\value{
A list containing Mean, Variance, and Standard Deviation
}
\description{
Calculates the mean and standard deviation from grouped data.
Accepts either a raw numeric vector or a frequency data frame.
}
