% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_describe}
\alias{group_describe}
\title{Comprehensive Grouped Description}
\usage{
group_describe(x, width = NULL)
}
\arguments{
\item{x}{Numeric vector OR frequency table}

\item{width}{Class interval width (optional, only used if x is raw data)}
}
\value{
A detailed list of descriptive statistics
}
\description{
Generates all descriptive statistics for a dataset.
}
