% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{group_exclusive_freq}
\alias{group_exclusive_freq}
\title{Grouped Exclusive Frequency Distribution}
\usage{
group_exclusive_freq(x, width = NULL)
}
\arguments{
\item{x}{Numeric vector (raw data)}

\item{width}{Class interval width (optional)}
}
\value{
Data frame
}
\description{
Creates a frequency table with right-open intervals \verb{[a, b)}.
Ideal for continuous data.
}
