% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_five_number}
\alias{group_five_number}
\title{Grouped Five Number Summary}
\usage{
group_five_number(x, width = NULL, type = c("exclusive", "inclusive"))
}
\arguments{
\item{x}{Numeric vector OR frequency table}

\item{width}{Class interval width (optional)}

\item{type}{Grouping method: "exclusive" or "inclusive"}
}
\value{
Named numeric vector
}
\description{
Grouped Five Number Summary
}
