% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catgorical.R
\name{group_frequency_table}
\alias{group_frequency_table}
\title{Generate a Professional Grouped Frequency Distribution Table}
\usage{
group_frequency_table(data, num_var, cat_var, bins = 5, width = NULL)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{num_var}{The numerical variable.}

\item{cat_var}{The categorical variable.}

\item{bins}{Integer. Number of bins.}

\item{width}{Numeric. Exact class width.}
}
\value{
A list containing the clean data and the gt table.
}
\description{
Generate a Professional Grouped Frequency Distribution Table
}
