% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_data.R
\name{group_inclusive_freq}
\alias{group_inclusive_freq}
\title{Grouped Inclusive Frequency Distribution}
\usage{
group_inclusive_freq(x, width = NULL)
}
\arguments{
\item{x}{Numeric vector (raw data)}

\item{width}{Class interval width (optional)}
}
\value{
Data frame
}
\description{
Creates a frequency table with closed intervals \verb{[a, b]}.
Ideal for discrete/integer data.
}
