% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_kurtosis}
\alias{group_kurtosis}
\title{Grouped Kurtosis}
\usage{
group_kurtosis(x)
}
\arguments{
\item{x}{Numeric vector OR frequency table}
}
\value{
excess kurtosis
}
\description{
Grouped Kurtosis
}
