% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_mean}
\alias{group_mean}
\title{Grouped Mean}
\usage{
group_mean(x)
}
\arguments{
\item{x}{Numeric vector OR frequency table}
}
\value{
Numeric mean
}
\description{
Computes mean from grouped frequency distribution or raw data.
}
