% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_median}
\alias{group_median}
\title{Grouped Median}
\usage{
group_median(x)
}
\arguments{
\item{x}{Numeric vector OR frequency table}
}
\value{
numeric median
}
\description{
Estimates median from grouped data.
}
