% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_mode}
\alias{group_mode}
\title{Grouped Mode}
\usage{
group_mode(x)
}
\arguments{
\item{x}{Numeric vector OR frequency table}
}
\value{
numeric mode
}
\description{
Estimates mode from grouped data.
}
