% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{group_ogive}
\alias{group_ogive}
\title{Ogive Curve}
\usage{
group_ogive(x, width = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{width}{optional interval width}
}
\value{
A ggplot object
}
\description{
Creates a modern cumulative frequency curve.
}
