% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_outliers}
\alias{group_outliers}
\title{Grouped Outliers}
\usage{
group_outliers(x)
}
\arguments{
\item{x}{Numeric vector OR frequency table}
}
\value{
Vector of outlying midpoints
}
\description{
Detects outliers using mean ± 2SD rule based on midpoints.
}
