% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_percentile}
\alias{group_percentile}
\title{Grouped Percentile (Helper for Median and Quartiles)}
\usage{
group_percentile(x, p)
}
\arguments{
\item{x}{Numeric vector OR frequency table}

\item{p}{Percentile to calculate (0 to 100)}
}
\value{
Numeric percentile value
}
\description{
Calculates any percentile for grouped data.
}
