% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{group_polygon}
\alias{group_polygon}
\title{Frequency Polygon}
\usage{
group_polygon(x, width = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{width}{optional interval width}
}
\value{
A ggplot object
}
\description{
Frequency Polygon
}
