% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_sd}
\alias{group_sd}
\title{Grouped Standard Deviation}
\usage{
group_sd(x, width = NULL, type = c("exclusive", "inclusive"))
}
\arguments{
\item{x}{Numeric vector OR frequency table}

\item{width}{Class interval width (optional)}

\item{type}{Grouping method: "exclusive" or "inclusive"}
}
\value{
numeric sd
}
\description{
Grouped Standard Deviation
}
