% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_skewness}
\alias{group_skewness}
\title{Grouped Skewness}
\usage{
group_skewness(x)
}
\arguments{
\item{x}{Numeric vector OR frequency table}
}
\value{
skewness
}
\description{
Grouped Skewness
}
