% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_variance}
\alias{group_variance}
\title{Grouped Variance}
\usage{
group_variance(x)
}
\arguments{
\item{x}{Numeric vector OR frequency table}
}
\value{
numeric variance
}
\description{
Computes sample variance from grouped data.
}
