% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pareto_chart}
\alias{pareto_chart}
\title{Professional Pareto Chart}
\usage{
pareto_chart(
  x,
  title = "Pareto Chart",
  xlab = "Category",
  ylab = "Frequency",
  fill_color = "#34495E",
  line_color = "#E74C3C"
)
}
\arguments{
\item{x}{categorical variable}

\item{title}{custom title}

\item{xlab}{custom x axis}

\item{ylab}{custom y axis}

\item{fill_color}{custom bar color}

\item{line_color}{custom line color}
}
\value{
A ggplot object
}
\description{
Professional Pareto Chart
}
