% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students-data.R
\docType{data}
\name{students}
\alias{students}
\title{Students Dataset}
\format{
A data frame with 85 rows and 2 variables:
\describe{
\item{age}{Age of student}
\item{gender}{Gender of student}
}
}
\source{
Created for statusluxR package
}
\usage{
students
}
\description{
A dataset containing age, and gender of students.
}
\keyword{datasets}
