% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cat_pie}
\alias{cat_pie}
\title{Pie Chart for Categorical Data}
\usage{
cat_pie(x)
}
\arguments{
\item{x}{categorical variable (vector)}
}
\value{
A ggplot object
}
\description{
Pie Chart for Categorical Data
}
