% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{data_diagnostics}
\alias{data_diagnostics}
\title{Data Cleansing and Diagnostic Summary}
\usage{
data_diagnostics(df)
}
\arguments{
\item{df}{target dataframe}
}
\value{
A list containing the diagnostic summary and the cleaned dataframe
}
\description{
Detects duplicates, removes them, and provides a missing value breakdown.
}
