% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_five_number}
\alias{group_five_number}
\title{Grouped Five Number Summary}
\usage{
group_five_number(x)
}
\arguments{
\item{x}{Numeric vector OR frequency table}
}
\value{
Named numeric vector
}
\description{
Calculates Min, Q1, Median, Q3, and Max for grouped data.
}
