% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catgorical.R
\name{group_frequency_table}
\alias{group_frequency_table}
\title{Generate a Professional Grouped Frequency Distribution Table}
\usage{
group_frequency_table(data, num_var, cat_var, bins = 5, width = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{num_var}{A numerical variable.}

\item{cat_var}{A categorical variable.}

\item{bins}{Number of classes.}

\item{width}{Optional class width.}
}
\value{
A list containing:
\itemize{
\item data: processed grouped data
\item table: a gt table
}
}
\description{
Creates a grouped frequency distribution table of a numerical variable
stratified by a categorical variable.
}
