% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_kurtosis}
\alias{group_kurtosis}
\title{Grouped Kurtosis}
\usage{
group_kurtosis(x, width = NULL, type = c("exclusive", "inclusive"))
}
\arguments{
\item{x}{Numeric vector OR frequency table}

\item{width}{Class interval width (optional)}

\item{type}{Grouping method: "exclusive" or "inclusive"}
}
\value{
excess kurtosis
}
\description{
Grouped Kurtosis
}
