% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_mean}
\alias{group_mean}
\title{Grouped Mean}
\usage{
group_mean(x, width = NULL, type = c("exclusive", "inclusive"))
}
\arguments{
\item{x}{Numeric vector OR frequency table}

\item{width}{Class interval width (optional)}

\item{type}{Grouping method: "exclusive" or "inclusive"}
}
\value{
Numeric mean
}
\description{
Grouped Mean
}
