% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_median}
\alias{group_median}
\title{Grouped Median}
\usage{
group_median(x, width = NULL, type = c("exclusive", "inclusive"))
}
\arguments{
\item{x}{Numeric vector OR frequency table}

\item{width}{Class interval width (optional)}

\item{type}{Grouping method: "exclusive" or "inclusive"}
}
\value{
numeric median
}
\description{
Grouped Median
}
