% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_mode}
\alias{group_mode}
\title{Grouped Mode}
\usage{
group_mode(x, width = NULL, type = c("exclusive", "inclusive"))
}
\arguments{
\item{x}{Numeric vector OR frequency table}

\item{width}{Class interval width (optional)}

\item{type}{Grouping method: "exclusive" or "inclusive"}
}
\value{
numeric mode
}
\description{
Grouped Mode
}
