% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_percentile}
\alias{group_percentile}
\title{Grouped Percentile}
\usage{
group_percentile(x, p, width = NULL, type = c("exclusive", "inclusive"))
}
\arguments{
\item{x}{Numeric vector OR frequency table}

\item{p}{Percentile to calculate (0 to 100)}

\item{width}{Class interval width (optional)}

\item{type}{Grouping method: "exclusive" or "inclusive"}
}
\value{
Numeric percentile value
}
\description{
Grouped Percentile
}
