% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_sd}
\alias{group_sd}
\title{Grouped Standard Deviation}
\usage{
group_sd(x)
}
\arguments{
\item{x}{Numeric vector OR frequency table}
}
\value{
numeric sd
}
\description{
Computes standard deviation from grouped data.
}
