% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{group_skewness}
\alias{group_skewness}
\title{Grouped Skewness}
\usage{
group_skewness(x, width = NULL, type = c("exclusive", "inclusive"))
}
\arguments{
\item{x}{Numeric vector OR frequency table}

\item{width}{Class interval width (optional)}

\item{type}{Grouping method: "exclusive" or "inclusive"}
}
\value{
skewness
}
\description{
Grouped Skewness
}
